Rearrange the structure of some folders and files, and create a new header component
