+++ src/app/projects/page.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import Image from "next/image";
 import { Metadata } from "next";
+
+import HeaderWithNavBar from "@/components/navigation/header-with-navbar";
 import MaxWidthWrapper from "@/components/ui/max-width-wrapper";
-import NavBar from "@/components/ui/navbar";
 import { Button } from "@/components/ui/button";
 import ProjectsList from "@/components/projects/projects-list";
 import { Project } from "@/lib/interfaces";
@@ -16,12 +17,12 @@ export const metadata: Metadata = {
     "Explore a curated selection of projects created by Sandra Höst Kannerberg, see how I bring ideas to life through design and development.",
 };
 
-export default async function ProjectPage() {
+export default async function ProjectsPage() {
   const projectsData = data as Project[];
 
   return (
     <>
-      <NavBar />
+      <HeaderWithNavBar />
       <MaxWidthWrapper>
         <main className="grid grid-cols-1 md:grid-cols-12">
           <section className="col-span-12 sm:col-span-6">
@@ -32,7 +33,7 @@ export default async function ProjectPage() {
               Projects
             </h1>
             <a href="https://github.com/SandraHKannerberg/portfolio">
-              <Button className="rounded shadow uppercase mt-5 mb-10 cursor-pointer hover:bg-secondary hover:text-background animate-bounce">
+              <Button className="rounded shadow uppercase mt-5 mb-10 cursor-pointer hover:bg-secondary hover:text-background">
                 Portfolio source code
               </Button>
             </a>
